refactor(workflows): Extract execution strategies from executor.py (#50)
Split the monolithic executor.py (1069 lines) into a strategy pattern
with dedicated modules per execution mode. The executor now handles
only orchestration (~390 lines) while mode-specific logic lives in
strategies/.

New files:
- strategies/base.py — ExecutionStrategy ABC + shared make_title() helper
- strategies/single.py — SingleStrategy
- strategies/parallel.py — ParallelStrategy (with worktree isolation)
- strategies/iterative.py — IterativeStrategy
- strategies/adversarial.py — AdversarialStrategy
- strategies/dynamic.py — DynamicStrategy (discovery + parallel)
- tests/test_strategies.py — 17 tests for dispatch, title helper, and
  strategy execution (single, iterative, adversarial)
